refactor: artistic filters, bug fixes, and codebase cleanup
Artistic filters:
- Add PencilSketchFilter and CartoonFilter (cv2.pencilSketch / cv2.stylization)
  as AsyncVideoFilter subclasses so heavy computation runs off the GUI thread
- 42 tests; both support pipeline export via to_code()

Bug fixes:
- _MedianBase._initialize: seed _buffer with data instead of zeros, eliminating
  all-black first-frame output from MoMedian
- _MedianBase.reset: clear shape=None so the next add() re-initializes with a
  seeded buffer (same fix applies post-reset, not just on first use)
- MoMedian.add: return early after _initialize so the seeded _result is preserved
- SampleHold.get: return raw frame during accumulation so display stays live
- _NormalizeMixin.get: use float32 instead of float64, halving per-frame allocation

Naming consistency:
- QEdgeFilter: rename setLow/setHigh -> _setLow/_setHigh
- QROIFilter: rename setX/setY/setW/setH -> _setX/_setY/_setW/_setH

Test cleanup:
- Remove 6 _MedianBase-level tests from test_momedian.py already covered by
  test_median.py (shape, seed, order setter, mutate-input)
- Update reset test to expect None result instead of zero-filled array
- Add tests for SampleHold.get() during accumulation and before first add

Style:
- Add American English spelling rule to CLAUDE.md
- Remove stray editor backup files (QBlobFilter.py~, QROIFilter.py~, QThresholdFilter.py~)
- Fix trailing whitespace in QVideoFilter._connectSignals

+++ filters/normalize.py
@@ -92,7 +92,7 @@ class _NormalizeMixin:
         if self._fg is None:
             raise RuntimeError('get() called before add()')
         bg = super().get()
-        result = np.zeros_like(self._fg, dtype=float)
+        result = np.zeros_like(self._fg, dtype=np.float32)
         np.divide(self._fg, bg, out=result, where=(bg != 0))
         if self.scale:
             result = self.mean * result
